+++ app/models/news_article.rb
@@ -20,7 +20,15 @@ class NewsArticle < Node
     order("content ->> 'name' ASC")
   }
 
+  scope :by_section, -> (section) {
+    where(:section_id => section.id)
+  }
+
   validates_presence_of :section
+
+  def self.published_for_section(section)
+    NewsArticle.by_section(section).by_release_date.by_published_at.published
+  end
 
   def layout
     'news_article'